+++ src/module/data/actor/party.mjs
@@ -146,8 +146,14 @@ export default class PartyModel extends DrawSteelSystemModel {
    * @returns {Promise<DrawSteelTokenDocument[]>}    A promise that resolves to the created tokens.
    */
   async placeMembers() {
-    const config = { tokens: this.members.map(m => m.actor.prototypeToken) };
-    const data = await ds.canvas.placeables.tokens.TokenPlacement.place(config);
-    return ds.canvas.placeables.tokens.TokenPlacement.createTokens(data);
+    if (!canvas?.scene) {
+      const msg = _loc("DRAW_STEEL.Actor.party.NoScene");
+      ui.notifications.error(msg, { console: false });
+      throw new Error(msg);
+    }
+
+    const tokenPromises = this.members.map(m => m.actor.getTokenDocument({}, { parent: canvas.scene }));
+    const createData = await Promise.all(tokenPromises);
+    return canvas.tokens.placeTokens(createData.map(t => t.toObject()));
   }
 }