+++ CHANGELOG.md
@@ -40,9 +40,12 @@ This version of the system does not support v13, it is exclusively for Foundry v
 
 ### Deprecated
 
+- Deprecated `canvas.tokens.performTokenPlacement` in favor of `canvas.tokens.placeActor`. (#1742)
+
 ### Removed
 
 - Removed generic Burning, Frozen, Marked, Targeted statuses from the token HUD. You can use an Active Effect compendium to create generic and reusable statuses. (#1256)
+- Removed `ds.canvas.placeables.tokens.TokenPlacement` in favor of delegating to `canvas.tokens.placeTokens`. (#1742)
 
 ### Fixed